requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
chromadb
pypdf
python-dotenv
requests
streamlit